git add .! Deprecate Llama-based models git add .!
