+++ legacy/src/fabrique/generation.py
