+++ legacy/src/fabrique/models/common/cache.py
