+++ legacy/src/fabrique/models/common/embeddings.py
