+++ legacy/src/fabrique/models/common/norm.py
