+++ legacy/src/fabrique/models/common/utils.py
