+++ legacy/src/fabrique/models/llama/load_rules.py
