+++ legacy/src/fabrique/models/llm.py
