+++ legacy/src/fabrique/models/phi/load_rules.py
