+++ legacy/src/fabrique/models/qwen2/__init__.py
