+++ legacy/src/fabrique/models/qwen2/load_rules.py
