--- src/fabrique/__init__.py
@@ -1,3 +1,0 @@
-from fabrique.models.llm import LLM, ChatMessage
-
-__all__ = ["LLM", "ChatMessage"]
