+++ src/fabrique/loading.py
@@ -1,20 +1,16 @@
-import importlib
-import json
-import os
-import pkgutil
+import logging
 import re
 from dataclasses import dataclass
-from typing import Callable, Dict, List
+from typing import Callable
 
 import jax
-import safetensors.flax as st
 from flax import nnx
-from huggingface_hub import snapshot_download
-from tokenizers import Tokenizer
-from tqdm import tqdm
+from jax.sharding import NamedSharding
 
-from fabrique import models
-from fabrique.utils import set_nested_attr
+from fabrique.utils import get_by_path, set_by_path
+
+logger = logging.getLogger("fabrique")
+
 
 ###############################################################################
 #                                  RULES                                      #
@@ -32,96 +28,58 @@ IGNORE = RuleIgnore()
 
 
 @dataclass
-class ConversionRule:
-    safe_pattern: str
-    fab_pattern: str | RuleIgnore
+class LoadRule:
+    in_pattern: str
+    out_pattern: str | RuleIgnore
     converter: Callable | None = None
 
-    @property
-    def safe_regexp(self):
-        pat = self.safe_pattern
-        pat = pat.replace(".", "\\.")
-        pat = pat.replace("{i}", "(?P<i>\\d+)")
-        pat = pat.replace("{j}", "(?P<j>\\d+)")
-        pat = pat.replace("{k}", "(?P<k>\\d+)")
-        pat = pat.replace("{n}", "(?P<n>\\d+)")
-        pat = "^" + pat + "$"
-        return re.compile(pat)
+    def __iter__(self):
+        yield self.in_pattern
+        yield self.out_pattern
+        yield self.converter
 
 
-def maybe_apply_rule(rule: ConversionRule, safe_key: str, safe_val):
-    if m := re.match(rule.safe_regexp, safe_key):
-        if isinstance(rule.fab_pattern, RuleIgnore):
-            return "", IGNORE
-        path = rule.fab_pattern.format(**m.groupdict())
-        val = rule.converter(safe_val) if rule.converter else safe_val
-        return path, val
-    else:
-        return None
+R = LoadRule
 
 
-def convert_safetensor(rules: List[ConversionRule], safe_key: str, safe_val):
-    for rule in rules:
-        path_val = maybe_apply_rule(rule, safe_key, safe_val)
-        if path_val:
-            return path_val
-    if isinstance(safe_val, jax.Array):
-        val_error_str = f"The value is an array with shape={safe_val.shape} " + \
-            f"and dtype={safe_val.dtype}"
-    else:
-        val_error_str = f"The value is of type {type(safe_val)}"
-    raise ValueError(
-        f"Couldn't find a rule for safetensors key {safe_key}. {val_error_str}"
-    )
+def _pattern_to_regexp(pat: str) -> str:
+    pat = pat.replace(".", "\\.")
+    pat = pat.replace("{i}", "(?P<i>\\d+)")
+    pat = pat.replace("{j}", "(?P<j>\\d+)")
+    pat = pat.replace("{k}", "(?P<k>\\d+)")
+    pat = pat.replace("{n}", "(?P<n>\\d+)")
+    pat = "^" + pat + "$"
+    return pat
 
 
-def apply_rules(
-    module: nnx.Module, rules: List[ConversionRule], flat: Dict[str, jax.Array]
-):
-    for safe_key, safe_val in flat.items():
-        path, val = convert_safetensor(rules, safe_key, safe_val)
-        if val is IGNORE:
-            continue
-        elif isinstance(val, jax.Array):
-            # setting paramater
-            fab_keys = path.split(".")
-            fab_keys += ["value"]  # set to the .value field
-            set_nested_attr(module, fab_keys, val)
-        elif isinstance(val, dict):
-            # setting value as is
-            # used e.g. when filling ToNNX-wrapperd Linen modules
-            fab_keys = path.split(".")
-            set_nested_attr(module, fab_keys, val)
-
-
-###############################################################################
-#                              MODEL UPDATE                                   #
-###############################################################################
-
-
-def update_module_from_safe(
-    module: nnx.Module, rules: List[ConversionRule], model_dir: str
-):
+def convert_path(path: str, in_pattern: str, out_pattern: str | RuleIgnore):
     """
-    Update Flax NNX model from a Huggingface model directory
+    Convert path according to input and output patterns. Example:
+
+    ```
+    path = "layer_3.attn.attn_vec_einsum.w"
+    in_pat = "layer_{n}.attn.attn_vec_einsum.w"
+    out_pat = "blocks.{n}.attn.attn_vec_einsum.kernel"
+
+    convert_path(path, in_pat, out_pat)
+    # ==> 'blocks.{n}.attn.attn_vec_einsum.kernel'
+    ```
+
     """
-    index_file = os.path.join(model_dir, "model.safetensors.index.json")
-    model_file = os.path.join(model_dir, "model.safetensors")
-    if os.path.isfile(index_file):
-        with open(index_file) as fp:
-            index = json.load(fp)
-        safe_files_ = set(index["weight_map"].values())
-        safe_files = [os.path.join(model_dir, filename) for filename in safe_files_]
-    elif os.path.isfile(model_file):
-        safe_files = [model_file]
-    else:
-        raise ValueError(f"Can't find safetensor files in {model_dir}")
-    for path in tqdm(safe_files):
-        flat = st.load_file(path)
-        apply_rules(module, rules, flat)
+    pat_re = _pattern_to_regexp(in_pattern)
+    if m := re.match(pat_re, path):
+        if isinstance(out_pattern, RuleIgnore):
+            return IGNORE
+        return out_pattern.format(**m.groupdict())
 
 
-def update_module_from_params(module: nnx.Module, rules: List[ConversionRule], params: dict):
+def update_module_from_params(
+    module: nnx.Module,
+    rules: tuple[str, str],
+    params: dict,
+    *,
+    mesh: jax.sharding.Mesh | None = None,
+):
     """
     Update Flax NNX module from a Flax Linen param tree
     """
@@ -129,92 +87,48 @@ def update_module_from_params(module: nnx.Module, rules: List[ConversionRule], p
     def keys_to_path(keys):
         return ".".join(key.key for key in keys)
 
-    state = nnx.state(module)                   # the model's state, a pure pytree
-    pspecs = nnx.get_partition_spec(state)      # strip out the annotations from state
-    # TODO: use mesh context manager (from outside?)
-    # TODO: move to apply_rules or somewhere to first map path in params and in model
-    # TODO: or maybe rework the whole loading mechanism
-    sharded_params = jax.lax.with_sharding_constraint(params, pspecs.raw_mapping)
-    flat_with_keys, _ = jax.tree.flatten_with_path(sharded_params)
-    flat = {keys_to_path(key): val for key, val in flat_with_keys}
-    apply_rules(module, rules, flat)
+    state = nnx.state(module)  # the model's state, a pure pytree
+    pspecs = nnx.get_partition_spec(state)  # strip out the annotations from state
+    for param_keys, val in jax.tree.flatten_with_path(params)[0]:
+        param_path = keys_to_path(param_keys)
+        for in_pattern, out_pattern, converter in rules:
+            module_path = convert_path(param_path, in_pattern, out_pattern)
+            if not module_path:
+                continue
+            # path is rules points to Param, but here we work with Array values
+            module_path += ".value"
+            if converter:
+                val = converter(val)
+            if mesh:
+                pspec = get_by_path(pspecs.raw_mapping, module_path)
+                val = jax.lax.with_sharding_constraint(val, NamedSharding(mesh, pspec))
+            set_by_path(module, module_path, val)
 
 
-###############################################################################
-#                             HUGGINGFACE HUB                                 #
-###############################################################################
+def main():
+    import jax.numpy as jnp
+    import numpy as np
+    from gemma import gm
 
+    from fabrique.models.gemma.load_rules import RULES
+    from fabrique.models.gemma.modeling import Transformer
 
-@dataclass
-class LoadConfig:
-    model_types: List[str]
-    model_args_class: type
-    model_class: type
-    rules: List[ConversionRule]
-    chat_template: str | None = None
+    # in_path = "layer_3.attn.q_einsum.w"
+    # in_pattern = "layer_{n}.attn.q_einsum.w"
+    # out_pattern = "blocks.{n}.attn.q_einsum.kernel.value"
+    # out_path = convert_path(in_path, in_pattern, out_pattern)
 
+    config = gm.nn.Gemma3_4B.config
+    ckpt = gm.ckpts.CheckpointPath.GEMMA3_4B_IT
+    params = gm.ckpts.load_params(ckpt)
+    module = nnx.eval_shape(
+        lambda: Transformer(config, param_dtype=jnp.bfloat16, rngs=nnx.Rngs(0))
+    )
 
-def find_load_configs():
-    load_configs = {}
-    for module_info in pkgutil.iter_modules(
-        models.__path__, prefix=models.__name__ + "."
-    ):
-        module = importlib.import_module(module_info.name)
-        if hasattr(module, "LOAD_CONFIG"):
-            cfg = getattr(module, "LOAD_CONFIG")
-            for model_type in cfg.model_types:
-                load_configs[model_type] = cfg
-    return load_configs
+    rules = RULES
+    mesh = jax.sharding.Mesh(
+        devices=np.array(jax.devices())[None, :], axis_names=("data", "model")
+    )
+    update_module_from_params(module, rules, params, mesh=mesh)
 
-
-def get_load_config(model_type: str):
-    # collect load configs by traversing fabrique.models subpackages
-    ret = find_load_configs().get(model_type)
-    if ret is None:
-        raise ValueError(f"Cannot find load config for model type {model_type}")
-    return ret
-
-
-def tweak_model_args(model_args):
-    "Slightly modify model args for better user experience"
-    model_args = {k: v for k, v in model_args.items()}  # make a copy
-    if "dtype" in model_args and "param_dtype" not in model_args:
-        model_args["param_dtype"] = model_args["dtype"]
-    return model_args
-
-
-def from_pretrained(model_id: str, revision: str | None = None, **model_args):
-    """
-    Load a model from a Huggingface Hub.
-
-    Args:
-        repo_id (str): ID of a model repo in Hugginface Hub.
-        **model_args: Keyword arguments to overwrite defaults in ModArgsC().
-
-    Returns:
-        Tuple of (tokenizer, model, hf_config).
-    """
-    model_dir = snapshot_download(model_id, revision=revision, repo_type="model")
-    # load config
-    config_file = os.path.join(model_dir, "config.json")
-    with open(config_file) as fp:
-        hf_config = json.load(fp)
-    # load tokenizer
-    tokenizer_file = os.path.join(model_dir, "tokenizer.json")
-    tokenizer = Tokenizer.from_file(tokenizer_file)
-    # load model
-    model_type = hf_config["model_type"]
-    cfg = get_load_config(model_type)
-    model_args = tweak_model_args(model_args)
-    args = cfg.model_args_class.from_file(config_file, **model_args)
-    model = cfg.model_class(args)
-    update_module_from_safe(model, cfg.rules, model_dir)
-    # load chat template
-    # note: in huggingface/transformers, chat template is loaded into tokenizer,
-    # but we use huggingface/tokenizers, which are slightly different.
-    # thus, we load the template seprately and put to hf_config
-    tokenizer_config_file = os.path.join(model_dir, "tokenizer_config.json")
-    with open(tokenizer_config_file) as fp:
-        tok_config = json.load(fp)
-        hf_config["chat_template"] = tok_config.get("chat_template", cfg.chat_template)
-    return tokenizer, model, hf_config
+    # param_keys, val = jax.tree.flatten_with_path(params)[0][8]
