+++ src/fabrique/models/gemma/layers.py
@@ -2,8 +2,6 @@ import jax
 import jax.numpy as jnp
 from flax import nnx
 
-from fabrique.utils import print_var
-
 
 class GemmaRMSNorm(nnx.Module):
     """
